requirements.txt
flake8
pytest-selenium
azure-cli